Moves harvester codes into package for harvesters
